Bug Fix: app now remembers it's current fragment even after screen rotation.
